# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: BlazeOS API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /trades:
    get:
      operationId: listTrades
      tags: [trades]
      summary: List all trades
      parameters:
        - in: query
          name: symbol
          schema:
            type: string
        - in: query
          name: result
          schema:
            type: string
        - in: query
          name: session
          schema:
            type: string
      responses:
        "200":
          description: List of trades
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Trade"
    post:
      operationId: createTrade
      tags: [trades]
      summary: Log a new trade
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TradeInput"
      responses:
        "201":
          description: Created trade
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trade"
...
  /trades/{id}:
    get:
      operationId: getTrade
      tags: [trades]
      summary: Get a single trade
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Trade
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trade"
        "404":
          description: Not found
    patch:
      operationId: updateTrade
      tags: [trades]
      summary: Update a trade
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TradeUpdate"
      responses:
        "200":
          description: Updated trade
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trade"
    delete:
      operationId: deleteTrade
      tags: [trades]
      summary: Delete a trade
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── CLIENTS ────────────────────────────────────────────────
...
components:
  schemas:
...
    Trade:
      type: object
      required: [id, symbol, direction, entryPrice, exitPrice, riskPercent, lotSize, setupType, session, result, createdAt]
      properties:
        id: { type: integer }
        symbol: { type: string }
        direction: { type: string }
        entryPrice: { type: number }
        exitPrice: { type: number }
        riskPercent: { type: number }
        lotSize: { type: number }
        setupType: { type: string }
        session: { type: string }
        result: { type: string }
        notes: { type: ["string", "null"] }
        screenshotBefore: { type: ["string", "null"] }
        screenshotAfter: { type: ["string", "null"] }
        pnl: { type: ["number", "null"] }
        createdAt: { type: string, format: date-time }
    TradeInput:
      type: object
      required: [symbol, direction, entryPrice, exitPrice, riskPercent, lotSize, setupType, session, result]
      properties:
        symbol: { type: string }
        direction: { type: string }
        entryPrice: { type: number }
        exitPrice: { type: number }
        riskPercent: { type: number }
        lotSize: { type: number }
        setupType: { type: string }
        session: { type: string }
        result: { type: string }
        notes: { type: string }
        screenshotBefore: { type: string }
        screenshotAfter: { type: string }
        pnl: { type: number }
    TradeUpdate:
      type: object
      properties:
        symbol: { type: string }
        direction: { type: string }
        entryPrice: { type: number }
        exitPrice: { type: number }
        riskPercent: { type: number }
        lotSize: { type: number }
        setupType: { type: string }
        session: { type: string }
        result: { type: string }
        notes: { type: string }
        screenshotBefore: { type: string }
        screenshotAfter: { type: string }
        pnl: { type: number }
...