# lib/api-zod/src/generated/types/tradeUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BlazeOS API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TradeUpdate {
  symbol?: string;
  direction?: string;
  entryPrice?: number;
  exitPrice?: number;
  riskPercent?: number;
  lotSize?: number;
  setupType?: string;
  session?: string;
  result?: string;
  notes?: string;
  screenshotBefore?: string;
  screenshotAfter?: string;
  pnl?: number;
}